REFACT:Update file-structure
atom, interface 분리
