+++ fe/src/atoms/style/atom.color.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const darkMode = atom({
+  key: 'dark',
+  default: false,
+});
