--- fe/src/components/analytics/DataSummary.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const DataSummaryWrapper = styled.div``;
-
-function DataSummary() {
-  return (
-    <DataSummaryWrapper>
-      <p>This is data summary</p>
-    </DataSummaryWrapper>
-  );
-}
-
-export default DataSummary;
