--- fe/src/components/analytics/Goal.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Goal() {
-  return <div>Goal</div>;
-}
-
-export default Goal;
