+++ fe/src/components/analytics/Visualization.tsx
@@ -1,8 +1,9 @@
 import React, { useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
-import { IVisualizationData, visualizationDatas } from '../../atoms/atom';
 import Chart from './chart/Chart';
+import { visualizationDatas } from '../../atoms/analytics/atom.analytics';
+import { IVisualizationData } from '../../interface/analytics/interface.analytics';
 
 const VisualHeader = styled.h1`
   font-weight: bold;
