--- fe/src/components/footer/evaluate.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Evaluate() {
-  return <div>Evaluate</div>;
-}
-
-export default Evaluate;
