+++ fe/src/components/form/file_upload_form/GenerateSettings.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
-import { generateSettingsMinMaxValue } from '../../../atoms/atom';
+import { generateSettingsMinMaxValue } from '../../../atoms/analytics/atom.analytics';
 
 const GenerateSettingsWrapper = styled.div`
   position: fixed;
