--- fe/src/components/form/file_upload_form/SampleData.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const DataBtn = styled.div`
-  cursor: pointer;
-  border: 1px solid black;
-  padding: 1px 2px;
-  margin: 10px 0;
-`;
-
-function SampleData() {
-  return (
-    <>
-      <DataBtn>sample Data</DataBtn>
-    </>
-  );
-}
-
-export default SampleData;
