+++ fe/src/components/main/team/Team.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { useRecoilValue } from 'recoil';
-import { introduce } from '../../../atoms/atom';
 import { motion } from 'framer-motion/dist/framer-motion';
 import Introduce from './Introduce';
+import { introduce } from '../../../atoms/main/atom.team';
 
 const Section = styled.section`
   padding: 150px 70px;
